** New Feature (initial phase) - will assign a threshold for sts scores, and below that minimum treshold, keywords will be placed to a temporary placeholder (Ad Group: Temp_Storage). - Minimum threshold is set at 0.85 STS score.
*to be done:
next phase of this feature that compares all keywords within the placeholder and assign them to new ad groups based on a threshold (shortest keyword will be the ad group name).
this may be done on a separate script as it may require that all keywords in the placeholder are already set.